### 2019-07-22 #326 @vbaliga
Submission: workloopR
Submitting Author: Vikram B. Baliga (@vbaliga)
Repository:   https://github.com/vbaliga/workloopR
Version submitted:   1.0.0
Version accepted: TBD

---

```
Package: workloopR
Title: Analysis of Work Loops and Other Data from Muscle Physiology Experiments
Version: 1.0.0
    person("Vikram B.", "Baliga",
        email = "vbaliga87@gmail.com",
        role = c("cre", "aut"),
    person("Shreeram", "Senthivasan",
        email = "shreeramsenthi@gmail.com",
        role = "aut",
        comment = c(ORCID = "0000-0002-7118-9547")))
Author: Vikram B. Baliga [cre, aut] (<https://orcid.org/0000-0002-9367-8974>),
    Shreeram Senthivasan [aut] (<https://orcid.org/0000-0002-7118-9547>)
Maintainer: Vikram B. Baliga <vbaliga87@gmail.com>
Description: Functions for the import, transformation, and analysis of data from muscle physiology experiments. The work loop technique is used to evaluate the mechanical work and power output of muscle. Josephson (1985) <https://jeb.biologists.org/content/114/1/493> modernized the technique for application in comparative biomechanics. Although our initial motivation was to provide functions to analyze work loop experiment data, as we developed the package we incorporated the ability to analyze data from experiments that are often complementary to work loops. There are currently three supported experiment types: work loops, simple twitches, and tetanus trials. Data can be imported directly from .ddf files or via an object constructor function. Through either method, data can then be cleaned or transformed via methods typically used in studies of muscle physiology. Data can then be analyzed to determine the timing and magnitude of force development and relaxation (for isometric trials) or the magnitude of work, net power, and instantaneous power among other things (for work loops). Although we do not provide plotting functions, all resultant objects are designed to be friendly to visualization via either base-R plotting or 'tidyverse' functions.
BugReports:https://github.com/vbaliga/workloopR/issues
    pracma (>= 2.0.7),
    signal (>= 0.7-6)
License: GPL (>= 3) + file LICENSE
RoxygenNote: 6.1.1.9000

```

This package offers the ability to import data from specialized file formats (including .ddf) that contain data from muscle physiology experiments. Data can then be transformed, re-organized, and analyzed according to methods typically used in studies of muscle physiology.

Muscle physiologists or any other researchers interested in characterizing the properties of muscular tissue.

We know of no other packages in R that handle work loop and related data sets.
