### 2019-08-05 #326 @lmullen
### Editor checks:

---

#### Editor comments

@vbaliga Thank you for your submission. I am currently looking for suitable reviewers. The review will begin when we have both in place.

In the meantime, here are a few things that you can work on.

- [ ] Please run `spelling::spell_check_package()` in your package. I note a few genuine misspellings (and of course false positives).

Otherwise R CMD check looks very clean.

More TK.

---

Reviewers:
Due date:
