### 2019-08-05 #326 @vbaliga
@lmullen Thank you for your feedback.

We used the `spelling::spell_check_package()` and took care of the misspellings throughout the package. The remaining items returned by that function are all proper nouns, specialized terms, and other false positives. We also fixed a few formatting issues in the package documentation. All changes are available now in [the latest commit](https://github.com/vbaliga/workloopR/commit/51ca63341023a996fecaa6057109aa8c16029b51).

With respect to your editor checks, would you recommend that we deposit our package in a repository (e.g. figshare) after the review process? We are interested in submitting to JOSS and want to make sure we are taking the right steps. Please let me know if JOSS review is contingent on having the package in a repository.

Thanks!
