### 2019-08-06 #326 @lmullen
@vbaliga I generally think it is much easier to do the archival version of the package once the review is done. Doubtless the reviewers will make changes, and you can sync up the reviewed version, the CRAN version, and the figshare version if the package is accepted.
