### 2019-10-08 #326 @jromanowska
*Please check off boxes as applicable, and elaborate in the comments below.  Your review is not limited to these topics, as described in the reviewer guide*

    * The installation instruction could be expanded to point out that the default command to install from github will __not__ create vignettes. While this may not be needed for all the users, I was a bit skeptic when I wanted to view the vignettes, as instructed in README.
    * Very good vignettes. Two comments from me:
      + could the authors include links to the vignettes in README, where they refer to those?
      + please, order the vignettes differently, so as to at least include "Intro..." as the first one.

##### For packages co-submitting to JOSS

- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)

    * There were no performance claims.
    * wrong use of `cat` in `data_import_functions.R` (this should be `message`)
    * README does not contain a list of vignettes, nor it has links to the vignettes
    * Documentation: nicely created with pkgdown; I would create groups for listing the functions in Reference

---

It is a very nice package, with a clear aim, providing specific functions for analysis of data from muscle physiology experiments. Even though I am not within the field, I could easily understand what it does and how to use it. The README is very instructive, nicely written, with clear figures and a fantastic workflow scheme to help explain what's going on in the package. Very good!

In general, the authors made a very good job of structuring the code, commenting, creating documentation and testing. However, I have a few comments and suggestions:

* is it necessary to have to install all the packages that are listed in vignettes? It takes a lot of time when creating those during the installation of the package.
* in the vignette "Plotting data in workloopR", please delete the `vnudge` parameter in the last command, as it produces a warning.
* why do the authors comment out some commands in the "examples" sections? Couldn't the "dontrun" command be used?
* I would suggest splitting the code into more files --- right now, there are three .R files with all the functions, which can be a bit difficult to read and maintain, especially if someone else would like to contribute
* please check that you use `return` command within each function to explicitly state what is the output of those
* `read_wl.ddf` and `read_twitch.ddf` mix the naming conventions --- I would suggest rather `read_wl_ddf` and `read_twitch_ddf`
* running `spelling::spell_check_package()` on the package showed that perhaps the authors should change "filetype", "filepath" to "file type" and "file path", respectively
* running `goodpractice::gp()` showed that there three things the authors should consider correcting:
  * too long lines in some of the functions
  * using both, "<-" and "=" is discouraged
  * change `sapply` to `vapply`, change `1:length()` to `seq_along`

Check also the [output](https://jromanowska.github.io/workloopR_review/) generated with the `pkgreviewr` for details on testing and automated checks.
