### 2019-10-08 #326 @eebrown
There is no conflict, but I note a caveat that this field is outside of my domain and I am doing this review from a technical perspective, as I understand it was hard to find reviewers within the domain with R knowledge.

This package is designed to analyze data from a specific set of muscle physiology experiments. More specifically, it analyzes the output from DDF files generated from a 3rd party proprietary device. The package’s features are explained in the README as well as the DESCRIPTION.

Yes, installation was straightforward.

The README provides a good overview of the features and available functions. There are several useful vignettes for the package. Given the work that has gone into creating these, it would be nice to have hyperlinks from the README page to access the rendered pages from the browser (the pkgdown website).

Functions are all documented as expected. Functions do not all have explicit return statements which is allowed but not ideal and makes code review harder in my opinion. I am surprised I don't see an @ropensci policy on this, so I'll leave it up to the authors/editor.

Please add contributor guidelines (esp. as you request particular types of contributors, e.g. to add support for non DDF file formats). e.g. https://devguide.ropensci.org/collaboration.html

##### For packages co-submitting to JOSS

- [X] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)

The package has full test coverage, including the expected results. However, it does not explain how the authors obtained the expected results for the tests.

It may be helpful to briefly explain the approach that was used for these analyses prior to this package (e.g. in the use case, what this is replacing). Further, it could be noted that (whether) the package produces the same results as this other method (was it some other software, was it manually). That would help prove the package functions as it claims, particularly since I don't have access to data to use to test the package for accuracy.

Are the default parameters (e.g. those in select_cycles) safe? I.e. would it be catastrophic if a user overlooked those arguments and didn't realize the defaults were in place? If so, it may be better not to have default arguments, and add tests to ensure the given arguments are possible values (rather than just numeric or not).

Passes the basic devtools::check().

Please add the rOpenSci review badge to the README (https://devguide.ropensci.org/building.html#readme)

In your DESCRIPTION file, the Author and Maintainer fields should not be included as they are redundant (they are generated from the Authors@R field).

_Are there improvements that could be made to the code style?_

A minor point is that the authors don't use the recommended spacing, e.g. they write `rising<-rising/100` instead of `rising <- rising / 100`. For these kinds of minor changes, consider using the `styler` package (https://github.com/r-lib/styler), as recommended (https://devguide.ropensci.org/building.html#code-style).

Running goodpractice::gp() provides the following suggestions which can be addressed:

>It is good practice to
>
>  ✖ use '<-' for assignment instead of '='. '<-' is the
>    standard, and R users and developers are used it and it is easier
>    to read your code for them if you use '<-'.
>
 >   R/data_analysis_functions.R:61:4
 >   R/data_analysis_functions.R:63:11
 >   R/data_analysis_functions.R:476:13
 >   R/data_import_functions.R:275:11
>
>  ✖ avoid long code lines, it is bad for readability. Also,
>   many people prefer editor windows that are about 80 characters
>    wide. Try make your lines shorter than 80 characters
>
>    R/data_analysis_functions.R:132:1
>    R/data_analysis_functions.R:224:1
>    R/data_analysis_functions.R:228:1
>    R/data_analysis_functions.R:249:1
>    R/data_analysis_functions.R:250:1
>    ... and 98 more lines
>
>  ✖ avoid sapply(), it is not type safe. It might return a
>    vector, or a list, depending on the input data. Consider using
>    vapply() instead.
>
>    R/data_analysis_functions.R:506:15
>    R/data_analysis_functions.R:507:17
>    R/data_analysis_functions.R:516:7
>    R/data_analysis_functions.R:518:7
>    R/data_import_functions.R:582:26
>    ... and 24 more lines
>
>  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...),
>    1:NROW(...) and 1:NCOL(...) expressions. They are error prone and
>    result 1:0 if the expression on the right hand side is zero. Use
>    seq_len() or seq_along() instead.
>
>    R/data_import_functions.R:275:13
>    R/data_import_functions.R:605:32

Please also note the typo on line 314 of data_import_functions.R.

_Is there code duplication in the package that should be reduced?_

You may wish to replace the internal function `trapezoidal_integration()` with `pracma::trapz()`, as you already import pracma, which is a more established package, and it would give you less to test/maintain.

_Are there user interface improvements that could be made?_

Linking to the documentation website from the repository/README could be helpful.

_Are there performance improvements that could be made?_

Appears fine to me.

Instead of just warning people that the units are assumed to be seconds, millimeters, and millinewtons, would it also be helpful to test that the data are in a reasonable range expected for those units? I assume there is a reasonable range for all muscle physiology experiments.

_Is the documentation (installation instructions/vignettes/examples/demos) clear and sufficient?_

Yes.

_Does it use the principle of multiple points of entry i.e. takes into account the fact that any piece of documentation may be the first encounter the user has with the package and/or the tool/data it wraps?_

Yes it appears sufficient and detailed, although it would be best to have this assessed by the target audience.

_Were functions and arguments named to work together to form a common, logical programming API that is easy to read, and autocomplete?_

Function names are logical and in families.

_If you have your own relevant data/problem, work through it with the package. You may find rough edges and use-cases the author didn’t think about._

Are there any DDF files from other sources (e.g. online repositories, published articles) that could be used to test?

---

My comments/suggestions are interspersed above. Overall I think the premise and design is excellent. On the face of it, this package seems very useful. My main concern as a reviewer from outside of the domain of muscle physiology, is that I cannot test it for accuracy. Please let me know if there is any way we can demonstrate this package provides a similar result to an existing analysis work flow, or commercial software, or previously done/published manual work, etc.
