### 2019-10-17 #326 @vbaliga
@jromanowska and @eebrown  Thank you both very much for your reviews. @shreeramsenthi and I are now going to work on the issues you both raised. We'll likely do this in several stages and spaced across a few commits. We'll notify you via this thread when we think we have responded to all your feedback and if we have any follow-up questions. Thanks again!
