### 2019-10-23 #326 @vbaliga
Hi @jromanowska and @eebrown,

Thank you both very much for your thorough reviews of workloopR. @shreeramsenthi and I feel that we have tackled most of the items you brought up. Please feel free to let us know if we can make any further improvements.

Also, as noted below, we took the liberty to list both of you as reviewers in our DESCRIPTION file. Please let us know if you'd prefer to not be listed or if you'd like any details changed.

Best regards,
Vikram

## Response to Package Review by @jromanowska

#### Documentation

  >- [x] **Installation instructions**: The installation instruction could be expanded to point out that the default command to install from github will **not** create vignettes. While this may not be needed for all the users, I was a bit skeptic when I wanted to view the vignettes, as instructed in README.

   **Response** We have edited the README with a note about building vignettes along with code in https://github.com/vbaliga/workloopR/commit/002ec083f5a708d520020e38d94b215562c23a81

>- [x] **Vignette(s)**: Very good vignettes. Two comments from me:
>    * could the authors include links to the vignettes in README, where they refer to those?
>    * please, order the vignettes differently, so as to at least include "Intro..." as the first one.

  **Response** We have added links to the vignettes within the README and reordered the vignette list on the pkgdown site in https://github.com/vbaliga/workloopR/commit/77e563c88b4dee51372fc0263decc7b458b0537d

#### Functionality

>- [ ] **Performance**: There were no performance claims

  **Response** Indeed, we make no performance claims as speed of execution does not seem to be a concern right now. Is this OK?

>- [ ] **Packaging guidelines**:
>   - wrong use of `cat()` in `data_import_functions.R` (this should be message)

  **Response** Could we please have clarification? `cat()` is used in three situations in `data_import_functions.R`:
    1. `print` methods for workloopR-specific S3 objects (muscle_stim, tetanus, etc.)
    2. `summary` methods for workloopR-specific S3 objects
    3. `print_muscle_stim_header()`, a non-exported function that is only ever called by the various print methods

  We see that points ii and iii do not fall strictly under the [rOpenSci guidelines](https://devguide.ropensci.org/building.html#console-messages) for using `cat()`, but they seem to fit a similar use case of displaying info from an object, rather than signaling condition like most `message()` calls we've come across. Would this use be OK?

>- [ ] **Packaging guidelines**:
>   - README does not contain a list of vignettes, nor it has links to the vignettes

  **Response** As noted above, we added links to the vignettes within the README in https://github.com/vbaliga/workloopR/commit/77e563c88b4dee51372fc0263decc7b458b0537d

>- [ ] **Packaging guidelines**:
>   - Documentation: nicely created with pkgdown; I would create groups for listing the functions in Reference

  **Response** This is a great suggestion and we were happy to comply: https://github.com/vbaliga/workloopR/commit/8f0de29d6f6523df98ad70eff0e2ec33571acc6a

#### Review Comments

>- is it necessary to have to install all the packages that are listed in vignettes? It takes a lot of time when creating those during the installation of the package.

**Response** Although the package does not depend on any `tidyverse` packages, the functions were written to fit into a `tidyverse`-centric workflow with minimal friction. Accordingly, we feel it is important to demonstrate this in our vignettes, which requires using functions from a variety of `tidyverse` packages. All other dependencies were removed https://github.com/vbaliga/workloopR/commit/b2abd81dc94cfae89b297a84e4e586d936bef5ee

> - in the vignette "Plotting data in workloopR", please delete the vnudge parameter in the last command, as it produces a warning.

**Response** The vnudge parameter has been deleted: https://github.com/vbaliga/workloopR/commit/77e563c88b4dee51372fc0263decc7b458b0537d

> - why do the authors comment out some commands in the "examples" sections? Couldn't the "dontrun" command be used?

**Response** Thank you for this suggestion. We swapped all such cases to use `dontrun()` in: https://github.com/vbaliga/workloopR/commit/a26cc7a36b6c740b9604a7ee7e94d1dace0488ab

> - I would suggest splitting the code into more files --- right now, there are three .R files with all the functions, which can be a bit difficult to read and maintain, especially if someone else would like to contribute

**Response** We agree and split the code into three additional files:  https://github.com/vbaliga/workloopR/commit/07cfdeff3c3597e4ed84ea64d279fbde922712d3

> - please check that you use return command within each function to explicitly state what is the output of those

**Response** We ensured `return()` was used in: https://github.com/vbaliga/workloopR/commit/4943ecc7c9f513949defda20d8d7defa77513120

> - read_wl.ddf and read_twitch.ddf mix the naming conventions --- I would suggest rather read_wl_ddf and read_twitch_ddf

**Response** We agree and made the change here: https://github.com/vbaliga/workloopR/commit/77e563c88b4dee51372fc0263decc7b458b0537d

> - running spelling::spell_check_package() on the package showed that perhaps the authors should change "filetype", "filepath" to "file type" and "file path", respectively

**Response** Thank you for catching this. Changes were made here: https://github.com/vbaliga/workloopR/commit/4d7b48cf89349fd33526b1dde857ec8bfd89166f

> - running goodpractice::gp() showed that there three things the authors should consider correcting:
>   - too long lines in some of the functions

**Response** We limited lines to 80 characters in: https://github.com/vbaliga/workloopR/commit/4d7b48cf89349fd33526b1dde857ec8bfd89166f

>    - using both, "<-" and "=" is discouraged

**Response** We agree and edited this in a few commits: https://github.com/vbaliga/workloopR/commit/56ee2c77b552f8cbc1792fca2bcecb65f59d64ff, and https://github.com/vbaliga/workloopR/commit/d654fec7c68d14068f68c25dfe47ec73c140f80f

>    - change sapply to vapply, change 1:length() to seq_along

**Response** We agree and also edited this across a few commits:  https://github.com/vbaliga/workloopR/commit/a1abf37018c67b9fc685be25a9e3aaca4ba6b38d, https://github.com/vbaliga/workloopR/commit/48b453b034bd59cb614cdfa93705a47be424924a, and https://github.com/vbaliga/workloopR/commit/5c74837234db4034aec0f2c7c33e9d17ab59d892

**Note** We both thank you for your helpful review. We would like to acknowledge your help and took the liberty of adding you as a reviewer in our DESCRIPTION file. Please let us know if you would not like to be acknowledged in this way or if you would like to censor or change any of your contact information.

-----

## Response to Package Review by @eebrown

#### Documentation

> - [x] **Vignette(s)**: The README provides a good overview of the features and available functions. There are several useful vignettes for the package. Given the work that has gone into creating these, it would be nice to have hyperlinks from the README page to access the rendered pages from the browser (the pkgdown website).

**Response** Thank you for your suggestion. We added links to the vignettes in the README in: https://github.com/vbaliga/workloopR/commit/77e563c88b4dee51372fc0263decc7b458b0537d

>- [x] **Function Documentation**: Functions are all documented as expected. Functions do not all have explicit return statements which is allowed but not ideal and makes code review harder in my opinion. I am surprised I don't see an @ropensci policy on this, so I'll leave it up to the authors/editor.

**Response** We agree and added `return()` statements to all named functions in: https://github.com/vbaliga/workloopR/commit/4943ecc7c9f513949defda20d8d7defa77513120

> - [ ] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with URL, BugReports and Maintainer (which may be autogeneated via Authors@R).
>  Please add contributor guidelines (esp. as you request particular types of contributors, e.g. to add support for non DDF file formats). e.g. https://devguide.ropensci.org/collaboration.html

**Response** Thank you for catching this. We were happy to add in contribution guidelines via `rodev::use_ro_github()` in: https://github.com/vbaliga/workloopR/commit/5cd03bd461d805486b920197aeb01175a4a1bb20

#### Functionality

> - [ ] Functionality: Any functional claims of the software been confirmed.

  > The package has full test coverage, including the expected results. However, it does not explain how the authors obtained the expected results for the tests.

  > It may be helpful to briefly explain the approach that was used for these analyses prior to this package (e.g. in the use case, what this is replacing). Further, it could be noted that (whether) the package produces the same results as this other method (was it some other software, was it manually). That would help prove the package functions as it claims, particularly since I don't have access to data to use to test the package for accuracy.

**Response** We agree with the points you make. Validity of the output is indeed something that should require due diligence. The approach that is typically used for these analyses is to use proprietary software such as Dynamic Muscle Analysis (DMA) by Aurora Software (see [here](https://aurorascientific.com/products/muscle-physiology/muscle-software/dynamic-muscle-control/)). The method for analysis by this and any other (valid) software follows what we show in Figure 1 of our JOSS paper: mechanical work is the product of force and distance and the area within a work loop is the net work produced by the muscle. Our own method of verification was to check the force, length, timing, net work reported by DMA against what our functions produce. We have done this for each of the example .ddf files included in workloopR. All values of force, timing, and net work are nearly identical to what DMA produces. At present, we are not sure the best way to show this to you and ropensci's editors. It is hard to generate an output file from DMA; instead values of e.g. net work are simply reported on-screen. Would supplying a variety of screenshots be sufficient? Happy to follow up as necessary

  > Are the default parameters (e.g. those in select_cycles) safe? I.e. would it be catastrophic if a user overlooked those arguments and didn't realize the defaults were in place? If so, it may be better not to have default arguments, and add tests to ensure the given arguments are possible values (rather than just numeric or not).

**Response** The default parameters in our functions conform to typical protocols in muscle physiology experiments, e.g. it is common to use cycles 3:5 in work loop studies. In that sense, our defaults should generally be safe. We also encourage users to think carefully about how their data are being analyzed and issue a variety of Warnings within many of the functions.

>- [ ] **Packaging guidelines** Please add the rOpenSci review badge to the README (https://devguide.ropensci.org/building.html#readme)

**Response** The badge has been added: https://github.com/vbaliga/workloopR/commit/a77eb6d3f566f48fb3e990d66e968fccde2d004f

>- [ ] **Packaging guidelines** In your DESCRIPTION file, the Author and Maintainer fields should not be included as they are redundant (they are generated from the Authors@R field).

**Response** The redundant Author and Maintainer fields have been removed in: https://github.com/vbaliga/workloopR/commit/b6184be6e275a2306ae7458e3cc3f41b39efd36d

>*Are there improvements that could be made to the code style?* A minor point is that the authors don't use the recommended spacing, e.g. they write rising<-rising/100 instead of rising <- rising / 100. For these kinds of minor changes, consider using the styler package (https://github.com/r-lib/styler), as recommended (https://devguide.ropensci.org/building.html#code-style).

**Response** Thanks for the suggestion. We used the `styler` package as recommended:  https://github.com/vbaliga/workloopR/commit/56ee2c77b552f8cbc1792fca2bcecb65f59d64ff and https://github.com/vbaliga/workloopR/commit/d654fec7c68d14068f68c25dfe47ec73c140f80f

>Running goodpractice::gp() provides the following suggestions which can be addressed:

**Response** Thanks for catching these. These have been addressed across a couple commits, most notably:
    - Removed sapply in: https://github.com/vbaliga/workloopR/commit/48b453b034bd59cb614cdfa93705a47be424924a and https://github.com/vbaliga/workloopR/commit/5c74837234db4034aec0f2c7c33e9d17ab59d892
    - Removed 1:length(...), etc. in: https://github.com/vbaliga/workloopR/commit/a1abf37018c67b9fc685be25a9e3aaca4ba6b38d
    - Broke up long lines in: https://github.com/vbaliga/workloopR/commit/f8c85babedd065f279985889368eab3843044913
    - Used '<-' consistently in: https://github.com/vbaliga/workloopR/commit/56ee2c77b552f8cbc1792fca2bcecb65f59d64ff

>Please also note the typo on line 314 of data_import_functions.R.

**Response** Thanks -- this has been fixed in: https://github.com/vbaliga/workloopR/commit/4ea5ea74c2439297e19796d5095f7155c68e3fa7

>*Is there code duplication in the package that should be reduced?* You may wish to replace the internal function trapezoidal_integration() with pracma::trapz(), as you already import pracma, which is a more established package, and it would give you less to test/maintain.

**Response** Although we see your point, we'd prefer to keep our function `trapezoidal_integration()`. We think it is a function that should be straightforward to test/maintain and keeping it in the package makes for one less thing to worry about if the authors of `pracma` choose to edit `trapz()`.

> *Are there user interface improvements that could be made?* Linking to the documentation website from the repository/README could be helpful.

**Response** Thanks for the suggestion; we added a link to the pkgdown site within the README in: https://github.com/vbaliga/workloopR/commit/77e563c88b4dee51372fc0263decc7b458b0537d

> Instead of just warning people that the units are assumed to be seconds, millimeters, and millinewtons, would it also be helpful to test that the data are in a reasonable range expected for those units? I assume there is a reasonable range for all muscle physiology experiments.

**Response** Although this would be nice to have, it is actually hard to predict what a reasonable range might be for a few of these units. Studies of muscle physiology range from examining minuscule flies to fish to large mammals and we hope this package will be useful to all such cases. Accordingly, it is hard to predict the scale of each variable as it will be strongly driven by organismal or muscular size.

> *If you have your own relevant data/problem, work through it with the package. You may find rough edges and use-cases the author didn’t think about.* Are there any DDF files from other sources (e.g. online repositories, published articles) that could be used to test?

**Response** A big source of frustration for us is the lack of open data in muscle physiology experiments in addition to the heavy reliance on proprietary software (hence, workloopR). The only other .ddf files that we know of in a public repository can be found in [this Figshare repository](https://figshare.com/articles/Data_analysis_from_Theriault_et_al_JEB/7336085/1). Please note that the study to which this repository corresponds is also from our laboratory (Altshuler Lab at UBC). So these .ddf files may be used to test our package, but please note there may be a potential conflict of interest (not exactly sure how this works).

> Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

**Response** Absolutely. We are very happy to do this and took the liberty to copy information about you that was available in the `chlorpromazineR` package DESCRIPTION file. Please let us know if you would like to change any of the information. Commit here: https://github.com/vbaliga/workloopR/commit/f0bad974ee7cb0739c4b5c038896ed6868ffc16b

Review Comments

> My comments/suggestions are interspersed above. Overall I think the premise and design is excellent. On the face of it, this package seems very useful. My main concern as a reviewer from outside of the domain of muscle physiology, is that I cannot test it for accuracy. Please let me know if there is any way we can demonstrate this package provides a similar result to an existing analysis work flow, or commercial software, or previously done/published manual work, etc.

**Response** Thanks again for all your feedback -- we found your review to be very helpful. As noted above, we're happy to follow up and show how our package produces similar results to those achieved in Aurora Software's Dynamic Muscle Analysis software (proprietary; see [here](https://aurorascientific.com/products/muscle-physiology/muscle-software/dynamic-muscle-control/)). If screenshots are sufficient, we can post them to this issue thread.
