### 2019-10-23 #326 @jromanowska
Great! Thank you for the very good response and description of the changes. A propos this "cat" usage - I can't find now any place that it's used wrongly. Maybe I've just misjudged, so you can forget about this.
I am very satisfied with the changes the authors made. Great work!
