### 2019-10-23 #326 @eebrown
Hi @vbaliga,
Thank you for addressing my feedback. I think your suggestion to show a screenshot of the commercial software output so that it could be compared side by side with your package's output would be a great asset this review as it would help me see how your results compare to the current  standard. I understand that there is not much open data available from other users to test at this time (and hopefully your package inspires more open research in this field!).
Otherwise, great work.
Thanks,
Eric
