### 2019-10-24 #326 @vbaliga
Hi @eebrown,

Please see [this document](https://github.com/ropensci/software-review/files/3769683/workloopR_vs_DMA.pdf), which we prepared to show the results of workloopR against those of commercial software. We showcase results using two types of experimental files. Let us know if you think what we've supplied is sufficient or if you have any follow-up questions.

Thanks again!
Vikram
