### 2019-10-25 #326 @eebrown
Dear @vbaliga,

Thank you very much for your thorough comparison of the package's output to an existing analysis suite. You clearly showed that your package produces the expected results with your supplied data files, and also illustrated the major benefits of your package in terms of replicability and open science. I appreciate the work you did to respond to this review and also hope that this comparison will be helpful to others considering using your package.

I have no further concerns and recommend this package for inclusion in rOpenSci. Congratulations and all the best.

Eric
