### 2019-10-25 #326 @vbaliga
Hi @eebrown,
Great! Thanks again for all your feedback. We appreciate your recommendation

@lmullen We believe that both reviewers are satisfied with the changes we made to workloopR. Please let us know if you have any feedback about the changes we made to the package, any comments on further changes that you'd like to see, and/or instructions for next steps

Thanks again!
Vikram
