### 2019-10-26 #326 @lmullen
@jromanowska @eebrown Thank you both for the thorough reviews and for going back and forth with the authors. Could you please both check off the relevant boxes, especially the ones about finally approving the package?

@vbaliga Thanks for the very clear responses to the reviewers work. Once they've checked off all the boxes we can work on next steps for moving this into rOpenSci.
