### 2019-10-26 #326 @jromanowska
Here is the final acceptance, after the authors replied to all the comments I had. Great work!

-----

*Please check off boxes as applicable, and elaborate in the comments below.  Your review is not limited to these topics, as described in the reviewer guide*

##### For packages co-submitting to JOSS

- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)

    * There were no performance claims.

- [x ] **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**
