### 2019-10-30 #326 @vbaliga
Thanks again @lmullen . I have now taken care of the remaining (Zenodo/JOSS) items.

The DOI is: 10.5281/zenodo.3523385
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3523385.svg)](https://doi.org/10.5281/zenodo.3523385)

* [X] Submit to JOSS at https://joss.theoj.org/papers/new, using the rOpenSci GitHub repo URL. When a JOSS "PRE REVIEW" issue is generated for your paper, add the comment: This package has been reviewed by rOpenSci: https://LINK.TO/THE/REVIEW/ISSUE, ropensci/editors

Please note that the JOSS PRE-REVIEW issue hasn't yet been generated. I'll be sure to follow up with posting the specified message when that happens.

Thanks again for all your help!
