### 2019-10-30 #326 @stefaniebutland
Hi @vbaliga. Congratulations on your package acceptance!

I think a tech note rather than a full length blog post on workloopR would be most appropriate since it has a very specific target audience. Instructions are here https://github.com/ropensci/roweb2#contributing-a-blog-post. Technically same as a blog post, but in YAML, set `categories: technotes` instead of `categories: blog`.

Your tech note could say what the package does and what unmet need it fills and provide a "cool" example use, without duplicating the vignette.

Please suggest a date on which you would like to submit a draft.
