### 2019-10-31 #326 @stefaniebutland
That sounds good. Please use `date: 2019-11-14` in YAML as the tentative publication date. Once submitted, I'll review 🙂
