### 2019-11-01 #326 @vbaliga
Hi @lmullen, @jromanowska, @eebrown ,

I am please to share that our paper has now been published in JOSS: https://doi.org/10.21105/joss.01856

Thanks again for all your help in reviewing workloopR and for suggesting improvements. @shreeramsenthi and I are both very happy with how everything turned out (but of course, we're happy to take further suggestions on features...etc).

Best regards,
Vikram
🐢
